refactor: make message builder use current session data service; remove unused files; clean up
