+++ src/routes/slack.ts
@@ -2,7 +2,6 @@ import express from 'express';
 import restrict from '../middleware/authorizeRole';
 import { getNewChannels } from '../controllers/slack/channel';
 import { getNewMembers } from '../controllers/slack/member';
-import { handleWeeklySessionsNotification } from '../controllers/slack/notification';
 import authMiddleware from '../middleware/authentication';
 import { receiver } from '../slack/slackApp';
 
@@ -22,18 +21,6 @@ router.get(
   restrict('admin'),
   getNewMembers
 );
-
-router.get(
-  '/weekly-sessions-notification',
-  express.json(),
-  authMiddleware,
-  handleWeeklySessionsNotification
-);
-
-router.use((req, res, next) => {
-  console.log('Incoming Request:', req.method, req.path, req.headers, req.body);
-  next();
-});
 
 // slack events, actions, commands
 router.use(receiver.router);
